db/explore/requirements.txt
psycopg2-binary>=2.9.3
tabulate>=0.8.10
PyYAML>=6.0
markdown>=3.3.6
toml>=0.10.2
